# PERF: improve performance of NDFrame.describe