A one-line change that enables to calculate the percentiles in `describe` more efficiently. The point is that calculating percentiles in one pass is faster than separately.

`describe` (with default `percentiles` argument) becomes 25-30% faster than before for numerical Series and DataFrames.

### Setup 
```
import timeit

setup = '''
import numpy as np
import pandas as pd
np.random.seed(123)
s = pd.Series(np.random.randint(0, 100, 1000000))
'''
```

### Benchmark
```
min(timeit.Timer('s.describe()', setup=setup).repeat(100, 1))
```

### Results
On master:
```
0.06349272100487724
```

With this change:
```
0.04745814300258644
```

Results are similar for DataFrames.

- [ ] closes #xxxx
- [x] tests added / passed
- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`
- [ ] whatsnew entry