@WillAyd 2018-05-31
Typically for performance-related changes we look for an ASV to measure and track over time. Can you add one to `asv_bench/benchmarks/frame_methods.py` and post the results of the benchmark here?