@WillAyd 2018-06-02
OK thanks. Can you update your PR to include the benchmark and a whatsnew note for 0.24?