@pep8speaks 2018-06-04
Hello @DataOmbudsman! Thanks for updating the PR.

 - In the file [`asv_bench/benchmarks/frame_methods.py`](https://github.com/pandas-dev/pandas/blob/70668a19291d00f0e970f2da6b91ee014b03e370/asv_bench/benchmarks/frame_methods.py), following are the PEP8 issues :

> [Line 516:1](https://github.com/pandas-dev/pandas/blob/70668a19291d00f0e970f2da6b91ee014b03e370/asv_bench/benchmarks/frame_methods.py#L516): [E302](https://duckduckgo.com/?q=pep8%20E302) expected 2 blank lines, found 1